requirements.txt
Django==2.2
django-rest-framework
factory_boy
requests

black
